All Cve tests combined to single file
