+++ Vulnerability.Scanner.Api/Tests/CVETests.cs
@@ -16,6 +16,18 @@ namespace Vulnerability.Scanner.Api.Tests
     [TestFixture]
     public class CVETests
     {
+        private static GroupScanResponse PutGroup(TestServer host, Guid groupId, IEnumerable<VulnerabilityScanRequest> scanRequests)
+        {
+            return host
+                .Put($"/v1/group/{groupId}", new GroupScanRequest
+                {
+                    ScanRequests = scanRequests
+                })
+                .ExpectStatusCode(HttpStatusCode.OK)
+                .WithContentOf<GroupScanResponse>()
+                .Select();
+        }
+
         [Test]
         public void WhenInvalidCVEIdIsFetched_ThenReturnNotFound()
         {
@@ -29,7 +41,7 @@ namespace Vulnerability.Scanner.Api.Tests
         }
 
         [Test]
-        public void WhenGroupIsAdded_ThenCVEHtmlPageShouldBeAvailable()
+        public void WhenGroupIsAdded_ThenCVESummaryShouldBeAvailable()
         {
             var host = SharedSetupFixture.Host;
             var groupId = Guid.NewGuid();
@@ -46,20 +58,28 @@ namespace Vulnerability.Scanner.Api.Tests
                 .Passing(x => x.Count().Should().BeGreaterThan(0));
         }
 
-        private static GroupScanResponse PutGroup(TestServer host, Guid groupId, IEnumerable<VulnerabilityScanRequest> scanRequests)
+        [Test]
+        public void WhenGroupIsCreated_ThenCheckCveCode()
         {
-            return host
-                .Put($"/v1/group/{groupId}", new GroupScanRequest
+            var host = SharedSetupFixture.Host;
+
+            var groupId = Guid.NewGuid();
+            var result = PutGroup(host, groupId, new [] { new VulnerabilityScanRequest("node:latest") });
+
+            host.Get(result.HtmlResultsUri)
+                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(15));
+
+            host.Get($"/v1/cve/")
+                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
+                .WithContentOf<IEnumerable<CveSummaryModel>>()
+                .Passing(x => 
                 {
-                    ScanRequests = scanRequests
-                })
-                .ExpectStatusCode(HttpStatusCode.OK)
-                .WithContentOf<GroupScanResponse>()
-                .Select();
+                   x.Should().Contain(a => a.Name == "CVE-2017-16997");
+                });
         }
 
         [Test]
-        public void WhenSelectedSortedByName_ThenCVEShouldBeSortedByName()
+        public void WhenSelectedSortedByName_ThenCVEShouldBeSortedRight()
         {
             var host = SharedSetupFixture.Host;
             var groupId = Guid.NewGuid();
@@ -73,16 +93,51 @@ namespace Vulnerability.Scanner.Api.Tests
             host.Get($"/v1/cve/?sortedBy=name")
                 .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
                 .WithContentOf<IEnumerable<CveSummaryModel>>()
-                .Passing(x => x.Should().BeInDescendingOrder(v => v.Name));
+                .Passing(x =>
+                {
+                    Assert.IsTrue(x.SequenceEqual( x.OrderByDescending(a => a.Name) ));
+                });
 
             host.Get($"/v1/cve/?sortedBy=nameDesc")
                 .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
                 .WithContentOf<IEnumerable<CveSummaryModel>>()
-                .Passing(x => x.Should().BeInAscendingOrder(v => v.Name));
+                .Passing(x =>
+                {
+                    Assert.IsTrue(x.SequenceEqual( x.OrderBy(a => a.Name) ));
+                });
         }
 
         [Test]
-        public void WhenSortedByInvalid_ThenCVEShouldBeSortedByDefault()
+        public void WhenSelectedSortedByFrequency_ThenCVEShouldBeSortedRight()
+        {
+            var host = SharedSetupFixture.Host;
+            var groupId = Guid.NewGuid();
+
+            var response = PutGroup(host, groupId, new [] { new VulnerabilityScanRequest("node:latest") });
+
+            host.Get(response.ResultsUri)
+                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
+                .WithContentOf<GroupScanResponse>();
+
+            host.Get($"/v1/cve/?sortedBy=frequency")
+                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
+                .WithContentOf<IEnumerable<CveSummaryModel>>()
+                .Passing(x =>
+                {
+                    Assert.IsTrue(x.SequenceEqual( x.OrderByDescending(a => a.Frequency) ));
+                });
+
+            host.Get($"/v1/cve/?sortedBy=frequencyDesc")
+                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
+                .WithContentOf<IEnumerable<CveSummaryModel>>()
+                .Passing(x =>
+                {
+                    Assert.IsTrue(x.SequenceEqual( x.OrderBy(a => a.Frequency) ));
+                });
+        }
+
+        [Test]
+        public void WhenSortedByInvalidValue_ThenCVEShouldBeSortedByDefault()
         {
             var host = SharedSetupFixture.Host;
             var groupId = Guid.NewGuid();
