--- Vulnerability.Scanner.Api/Tests/CveTest.cs
@@ -1,137 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using FluentAssertions;
-using Microsoft.AspNetCore.TestHost;
-using NUnit.Framework;
-using Protacon.NetCore.WebApi.TestUtil;
-using Protacon.NetCore.WebApi.TestUtil.Extensions;
-using Vulnerability.Scanner.Api.Controllers.Models;
-using Vulnerability.Scanner.Api.Data;
-using Vulnerability.Scanner.Api.Model;
-using Vulnerability.Scanner.Api.Tests;
-
-namespace Vulnerability.Scanner.Api.Tests
-{
-    [TestFixture]
-    public class CveTest
-    {
-
-         private static GroupScanResponse PutGroup(TestServer host, Guid groupId, IEnumerable<VulnerabilityScanRequest> scanRequests)
-        {
-            return host
-                .Put($"/v1/group/{groupId}", new GroupScanRequest
-                {
-                    ScanRequests = scanRequests
-                })
-                .ExpectStatusCode(HttpStatusCode.OK)
-                .WithContentOf<GroupScanResponse>()
-                .Select();
-        }
-
-
-        [Test]
-        public void WhenGroupIsCreated_ThenCheckCveCode()
-        {
-            var host = SharedSetupFixture.Host;
-
-            var groupId = Guid.NewGuid();
-            var result = PutGroup(host, groupId, new [] { new VulnerabilityScanRequest("node:latest") });
-
-            host.Get(result.HtmlResultsUri)
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(15));
-
-            host.Get($"/v1/cve/")
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
-                .WithContentOf<IEnumerable<CveSummaryModel>>()
-                .Passing(x => 
-                {
-                   x.Should().Contain(a => a.Name == "CVE-2017-16997");
-                });
-        }
-
-        [Test]
-        public void CheckIsNameSorted_ThenContinue()
-        {
-            var host = SharedSetupFixture.Host;
-
-            var groupId = Guid.NewGuid();
-            var result = PutGroup(host, groupId, new [] { new VulnerabilityScanRequest("node:latest") });
-
-            host.Get(result.HtmlResultsUri)
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(15));
-
-            host.Get($"/v1/cve?sortedBy=name")
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
-                .WithContentOf<IEnumerable<CveSummaryModel>>()
-                .Passing(x =>
-                {
-                    var expectedList = x.OrderByDescending(a => a.Name);
-                    Assert.IsTrue(expectedList.SequenceEqual(x));
-
-                });
-        }
-
-        [Test]
-        public void CheckIsNameDescSorted_ThenContinue()
-        {
-            var host = SharedSetupFixture.Host;
-
-            var groupId = Guid.NewGuid();
-            var result = PutGroup(host, groupId, new [] { new VulnerabilityScanRequest("node:latest") });
-
-            host.Get(result.HtmlResultsUri)
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(15));
-
-            host.Get($"/v1/cve?sortedBy=nameDesc")
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
-                .WithContentOf<IEnumerable<CveSummaryModel>>()
-                .Passing(x =>
-                {
-                    var expectedList = x.OrderBy(a => a.Name);
-                    Assert.IsTrue(expectedList.SequenceEqual(x));
-
-                });
-        }
-
-
-        [Test]
-        public void CheckIsFrequencySorted_ThenContinue()
-        {
-            var host = SharedSetupFixture.Host;
-
-            var groupId = Guid.NewGuid();
-            var result = PutGroup(host, groupId, new [] { new VulnerabilityScanRequest("node:latest") });
-
-            host.Get(result.HtmlResultsUri)
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(15));
-
-            host.Get($"/v1/cve?sortedBy=frequency")
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10))
-                .WithContentOf<IEnumerable<CveSummaryModel>>()
-                .Passing(x =>
-                {
-                    var expectedList = x.OrderByDescending(a => a.Frequency);
-                    Assert.IsTrue(expectedList.SequenceEqual(x));
-                    
-                });
-        }
-
-        [Test]
-        public void WhenCveIsAdded_ThenHtmlPageShouldBeAvailable()
-        {
-            var host = SharedSetupFixture.Host;
-            var groupId = Guid.NewGuid();
-
-            var response = PutGroup(host, groupId, new [] { new VulnerabilityScanRequest("node:latest") });
-
-            host.Get(response.HtmlResultsUri)
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10));
-
-            host.Get($"/v1/cve/html")
-                .WaitForStatusCode(HttpStatusCode.OK, timeout: TimeSpan.FromSeconds(10));
-        }
-
-    }
-}
